Display both round result and final result on page
As the user input is still by prompt,
only the round result for the final round
actually gets shown on the page at this point.

Round result and final result are displayed
in separate divs to allow separate styles
and for the final round result to be displayed
at the same time as the overall result.